Move src/cli.js logic into new src/lib/main.js file
